Xero/project/requirements.txt
pandas
sqlalchemy
psycopg2-binary
mcp
flask
flask-cors

